KASM-3747 Fix dnf reinstall removing pem cert

+++ oracle/kasmvncserver9.spec
@@ -154,4 +154,11 @@ cd $DST_MAN && ln -s vncpasswd.1 kasmvncpasswd.1;
   make_self_signed_certificate
 
 %postun
-  rm -f /etc/pki/tls/private/kasmvnc.pem
+  local is_uninstall = 0;
+
+  if [ "$1" == 0 ]; then
+    is_uninstall = 1;
+  fi
+  if [ "$is_uninstall" = 1 ]; then
+    rm -f /etc/pki/tls/private/kasmvnc.pem
+  fi
